chore: 요일별 애니 페이지 폴더 app/default로 이동 #
